200-other/205-ai-model-scanner/requirements.txt
httpx==0.28.1
supabase==2.13.0
python-dotenv==1.1.0
